+++ test/datalog_tests/server_api/tests/events.rs
@@ -286,7 +286,7 @@ fn setup() -> (DDlogServer, UpdatesObservable, MockObserver) {
     let mut server = DDlogServer::new(program, hashmap! {});
 
     let observer = SharedObserver::new(Mock::new());
-    let mut stream = server.add_stream(hashset! {server_api_1_P1Out});
+    let mut stream = server.add_stream(hashset! {server_api_1_P1Out as usize});
     let _ = stream.subscribe(Box::new(observer.clone())).unwrap();
 
     (server, stream, observer)
@@ -297,7 +297,7 @@ fn setup_tcp() -> (DDlogServer, UpdatesObservable, MockObserver, Box<dyn Any>) {
     let mut server = DDlogServer::new(program, hashmap! {});
 
     let observer = SharedObserver::new(Mock::new());
-    let mut stream = server.add_stream(hashset! {server_api_1_P1Out});
+    let mut stream = server.add_stream(hashset! {server_api_1_P1Out as usize});
 
     let mut recv = TcpReceiver::<Update<Value>>::new("127.0.0.1:0").unwrap();
     let send = TcpSender::connect(*recv.addr()).unwrap();
